initial gulp setup done
