+++ README.md
@@ -1,1 +1,4 @@
 This is a chrome extension that extends Pinterest's feature set by allowing the user to download images and videos.
+
+To generate dev version: gulp --watch
+To generate production version: gulp --config PROD
