+++ package.json
@@ -1,0 +1,24 @@
+{
+  "name": "pinterest_media_downloader",
+  "version": "1.0.0",
+  "description": "Chrome extension to download images and videos from Pinterest",
+  "main": "gulpfile.js",
+  "dependencies": {
+    "del": "^5.1.0",
+    "gulp": "^4.0.2",
+    "gulp-merge-json": "^2.1.0"
+  },
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/AbhinavS96/Pinterest-image-download-chrome-extension.git"
+  },
+  "author": "Abhinav",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/AbhinavS96/Pinterest-image-download-chrome-extension/issues"
+  },
+  "homepage": "https://github.com/AbhinavS96/Pinterest-image-download-chrome-extension#readme"
+}
