+++ src/config.PROD.json
@@ -1,0 +1,3 @@
+{
+  "title": "Hello, World!"
+}
