+++ src/manifest.PROD.json
@@ -1,0 +1,166 @@
+{
+  "manifest_version": 3,
+  "name": "Pinterest Video and Image Downloader",
+  "description": "This extension will help you download Pinterest pictures and videos to your PC",
+  "version": "5.4",
+  "icons": {
+    "128": "/assets/icon_128.png",
+    "48": "/assets/icon_48.png",
+    "16": "/assets/icon_16.png"
+  },
+  "action": {
+    "default_icon": {
+      "16": "/assets/icon_16.png"
+    },
+    "default_popup": "index.html",
+    "default_title": "Pinterest Video and Image Downloader"
+  },
+  "background": {
+    "service_worker": "eventPage.js"
+  },
+  "content_scripts": [
+    {
+      "matches": [
+        "*://*.pinterest.com/*",
+        "*://*.pinterest.ru/*",
+        "*://*.pinterest.de/*",
+        "*://*.pinterest.se/*",
+        "*://*.pinterest.pt/*",
+        "*://*.pinterest.at/*",
+        "*://*.pinterest.dk/*",
+        "*://*.pinterest.hu/*",
+        "*://*.pinterest.be/*",
+        "*://*.pinterest.ca/*",
+        "*://*.pinterest.in/*",
+        "*://*.pinterest.ec/*",
+        "*://*.pinterest.co/*",
+        "*://*.pinterest.cr/*",
+        "*://*.pinterest.id/*",
+        "*://*.pinterest.nz/*",
+        "*://*.pinterest.uk/*",
+        "*://*.pinterest.it/*",
+        "*://*.pinterest.info/*",
+        "*://*.pinterest.nl/*",
+        "*://*.pinterest.ch/*",
+        "*://*.pinterest.pe/*",
+        "*://*.pinterest.th/*",
+        "*://*.pinterest.tk/*",
+        "*://*.pinterest.jp/*",
+        "*://*.pinterest.ie/*",
+        "*://*.pinterest.cl/*",
+        "*://*.pinterest.tw/*",
+        "*://*.pinterest.es/*",
+        "*://*.pinterest.ph/*",
+        "*://*.pinterest.vn/*",
+        "*://*.pinterest.ag/*",
+        "*://*.pinterest.cz/*",
+        "*://*.pinterest.gs/*",
+        "*://*.pinterest.am/*",
+        "*://*.pinterest.bz/*",
+        "*://*.pinterest.cn/*",
+        "*://*.pinterest.mx/*",
+        "*://*.pinterest.co.nz/*",
+        "*://*.pinterest.co.at/*",
+        "*://*.pinterest.co.cr/*",
+        "*://*.pinterest.co.bo/*",
+        "*://*.pinterest.co.es/*",
+        "*://*.pinterest.co.py/*",
+        "*://*.pinterest.co.uk/*",
+        "*://*.pinterest.co.in/*",
+        "*://*.pinterest.co.vn/*",
+        "*://*.pinterest.co.pt/*",
+        "*://*.pinterest.co.pe/*",
+        "*://*.pinterest.com.bo/*",
+        "*://*.pinterest.com.co/*",
+        "*://*.pinterest.com.cn/*",
+        "*://*.pinterest.com.au/*",
+        "*://*.pinterest.com.ag/*",
+        "*://*.pinterest.com.es/*",
+        "*://*.pinterest.com.py/*",
+        "*://*.pinterest.com.ph/*",
+        "*://*.pinterest.com.uk/*",
+        "*://*.pinterest.com.vn/*",
+        "*://*.pinterest.com.pt/*",
+        "*://*.pinterest.com.pl/*",
+        "*://*.pinterest.com.pe/*",
+        "*://*.pinterest.com.tw/*",
+        "*://*.pinterest.com.ve/*",
+        "*://*.pinterest.com.mx/*"
+      ],
+      "js": [
+        "/scripts/content.js",
+        "/scripts/jszip.min.js",
+        "/scripts/FileSaver.min.js"
+      ],
+      "run_at": "document_end"
+    }
+  ],
+  "host_permissions": [
+    "*://*.pinterest.com/*",
+    "*://*.pinterest.ru/*",
+    "*://*.pinterest.de/*",
+    "*://*.pinterest.se/*",
+    "*://*.pinterest.pt/*",
+    "*://*.pinterest.at/*",
+    "*://*.pinterest.dk/*",
+    "*://*.pinterest.hu/*",
+    "*://*.pinterest.be/*",
+    "*://*.pinterest.ca/*",
+    "*://*.pinterest.in/*",
+    "*://*.pinterest.ec/*",
+    "*://*.pinterest.co/*",
+    "*://*.pinterest.cr/*",
+    "*://*.pinterest.id/*",
+    "*://*.pinterest.nz/*",
+    "*://*.pinterest.uk/*",
+    "*://*.pinterest.it/*",
+    "*://*.pinterest.info/*",
+    "*://*.pinterest.nl/*",
+    "*://*.pinterest.ch/*",
+    "*://*.pinterest.pe/*",
+    "*://*.pinterest.th/*",
+    "*://*.pinterest.tk/*",
+    "*://*.pinterest.jp/*",
+    "*://*.pinterest.ie/*",
+    "*://*.pinterest.cl/*",
+    "*://*.pinterest.tw/*",
+    "*://*.pinterest.es/*",
+    "*://*.pinterest.ph/*",
+    "*://*.pinterest.vn/*",
+    "*://*.pinterest.ag/*",
+    "*://*.pinterest.cz/*",
+    "*://*.pinterest.gs/*",
+    "*://*.pinterest.am/*",
+    "*://*.pinterest.bz/*",
+    "*://*.pinterest.cn/*",
+    "*://*.pinterest.mx/*",
+    "*://*.pinterest.co.nz/*",
+    "*://*.pinterest.co.at/*",
+    "*://*.pinterest.co.cr/*",
+    "*://*.pinterest.co.bo/*",
+    "*://*.pinterest.co.es/*",
+    "*://*.pinterest.co.py/*",
+    "*://*.pinterest.co.uk/*",
+    "*://*.pinterest.co.in/*",
+    "*://*.pinterest.co.vn/*",
+    "*://*.pinterest.co.pt/*",
+    "*://*.pinterest.co.pe/*",
+    "*://*.pinterest.com.bo/*",
+    "*://*.pinterest.com.co/*",
+    "*://*.pinterest.com.cn/*",
+    "*://*.pinterest.com.au/*",
+    "*://*.pinterest.com.ag/*",
+    "*://*.pinterest.com.es/*",
+    "*://*.pinterest.com.py/*",
+    "*://*.pinterest.com.ph/*",
+    "*://*.pinterest.com.uk/*",
+    "*://*.pinterest.com.vn/*",
+    "*://*.pinterest.com.pt/*",
+    "*://*.pinterest.com.pl/*",
+    "*://*.pinterest.com.pe/*",
+    "*://*.pinterest.com.tw/*",
+    "*://*.pinterest.com.ve/*",
+    "*://*.pinterest.com.mx/*"
+  ],
+  "incognito": "split"
+}
